Fix path handling in setup.py
Should correct #23 and #24

The '.amap' folder is created (by install_cfg) before starting to
download the atlas. The '/atlas' folder is created by the `tar`
command.
`amend_cfg` amends only the local config, not that in the code
(can be changed) and match the part of the name without 'atlas'